servers/parser/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
et-xmlfile==1.1.0
fastapi==0.115.2
h11==0.14.0
idna==3.10
motor==3.6.0
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
pydantic==2.9.2
pydantic-mongo==2.3.0
pydantic_core==2.23.4
pymongo==4.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.39.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
xlrd==2.0.1
aiobotocore==2.16.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioitertools==0.12.0
aiosignal==1.3.2
async-timeout==5.0.1
attrs==24.3.0
botocore==1.35.81
frozenlist==1.5.0
jmespath==1.0.1
multidict==6.1.0
propcache==0.2.1
wrapt==1.17.0
yarl==1.18.3